+++ index.ts
@@ -2,6 +2,7 @@ import { IncomingMessage, ServerResponse } from "http";
 import contentSecurityPolicy, {
   ContentSecurityPolicyOptions,
 } from "./middlewares/content-security-policy";
+import crossOriginEmbedderPolicy from "./middlewares/cross-origin-embedder-policy";
 import expectCt, { ExpectCtOptions } from "./middlewares/expect-ct";
 import originAgentCluster from "./middlewares/origin-agent-cluster";
 import referrerPolicy, {
@@ -26,6 +27,7 @@ import xXssProtection from "./middlewares/x-xss-protection";
 
 interface HelmetOptions {
   contentSecurityPolicy?: MiddlewareOption<ContentSecurityPolicyOptions>;
+  crossOriginEmbedderPolicy?: boolean;
   dnsPrefetchControl?: MiddlewareOption<XDnsPrefetchControlOptions>;
   expectCt?: MiddlewareOption<ExpectCtOptions>;
   frameguard?: MiddlewareOption<XFrameOptionsOptions>;
@@ -39,7 +41,7 @@ interface HelmetOptions {
   xssFilter?: MiddlewareOption<never>;
 }
 
-type MiddlewareOption<T> = false | T;
+type MiddlewareOption<T> = boolean | T;
 
 interface MiddlewareFunction {
   (
@@ -57,6 +59,7 @@ interface Helmet {
   ) => void;
 
   contentSecurityPolicy: typeof contentSecurityPolicy;
+  crossOriginEmbedderPolicy: typeof crossOriginEmbedderPolicy;
   dnsPrefetchControl: typeof xDnsPrefetchControl;
   expectCt: typeof expectCt;
   frameguard: typeof xFrameOptions;
@@ -64,14 +67,153 @@ interface Helmet {
   hsts: typeof strictTransportSecurity;
   ieNoOpen: typeof xDownloadOptions;
   noSniff: typeof xContentTypeOptions;
+  originAgentCluster: typeof originAgentCluster;
   permittedCrossDomainPolicies: typeof xPermittedCrossDomainPolicies;
   referrerPolicy: typeof referrerPolicy;
   xssFilter: typeof xXssProtection;
-  originAgentCluster: typeof originAgentCluster;
 
   featurePolicy: () => never;
   hpkp: () => never;
   noCache: () => never;
+}
+
+function getArgs<T>(
+  option: undefined | Readonly<MiddlewareOption<T>>,
+  middlewareConfig: Readonly<
+    {
+      enabledByDefault?: boolean;
+    } & (
+      | { takesOptions?: true }
+      | {
+          name: string;
+          takesOptions: false;
+        }
+    )
+  > = {}
+): null | [] | [T] {
+  const { enabledByDefault = true } = middlewareConfig;
+
+  switch (option) {
+    case undefined:
+      return enabledByDefault ? [] : null;
+    case false:
+      return null;
+    case true:
+      return [];
+    default:
+      if (middlewareConfig.takesOptions === false) {
+        console.warn(
+          `${middlewareConfig.name} does not take options. ${
+            enabledByDefault
+              ? "Remove the property"
+              : "Set the property to `true`"
+          } to silence this warning.`
+        );
+        return [];
+      } else {
+        return [option];
+      }
+  }
+}
+
+function getMiddlewareFunctionsFromOptions(
+  options: Readonly<HelmetOptions>
+): MiddlewareFunction[] {
+  const result: MiddlewareFunction[] = [];
+
+  const contentSecurityPolicyArgs = getArgs(options.contentSecurityPolicy);
+  if (contentSecurityPolicyArgs) {
+    result.push(contentSecurityPolicy(...contentSecurityPolicyArgs));
+  }
+
+  const crossOriginEmbedderPolicyArgs = getArgs(
+    options.crossOriginEmbedderPolicy,
+    {
+      name: "crossOriginEmbedderPolicy",
+      takesOptions: false,
+      enabledByDefault: false,
+    }
+  );
+  if (crossOriginEmbedderPolicyArgs) {
+    result.push(crossOriginEmbedderPolicy());
+  }
+
+  const xDnsPrefetchControlArgs = getArgs(options.dnsPrefetchControl);
+  if (xDnsPrefetchControlArgs) {
+    result.push(xDnsPrefetchControl(...xDnsPrefetchControlArgs));
+  }
+
+  const expectCtArgs = getArgs(options.expectCt);
+  if (expectCtArgs) {
+    result.push(expectCt(...expectCtArgs));
+  }
+
+  const xFrameOptionsArgs = getArgs(options.frameguard);
+  if (xFrameOptionsArgs) {
+    result.push(xFrameOptions(...xFrameOptionsArgs));
+  }
+
+  const xPoweredByArgs = getArgs(options.hidePoweredBy, {
+    name: "hidePoweredBy",
+    takesOptions: false,
+  });
+  if (xPoweredByArgs) {
+    result.push(xPoweredBy());
+  }
+
+  const strictTransportSecurityArgs = getArgs(options.hsts);
+  if (strictTransportSecurityArgs) {
+    result.push(strictTransportSecurity(...strictTransportSecurityArgs));
+  }
+
+  const xDownloadOptionsArgs = getArgs(options.ieNoOpen, {
+    name: "ieNoOpen",
+    takesOptions: false,
+  });
+  if (xDownloadOptionsArgs) {
+    result.push(xDownloadOptions());
+  }
+
+  const xContentTypeOptionsArgs = getArgs(options.noSniff, {
+    name: "noSniff",
+    takesOptions: false,
+  });
+  if (xContentTypeOptionsArgs) {
+    result.push(xContentTypeOptions());
+  }
+
+  const originAgentClusterArgs = getArgs(options.originAgentCluster, {
+    name: "originAgentCluster",
+    takesOptions: false,
+    enabledByDefault: false,
+  });
+  if (originAgentClusterArgs) {
+    result.push(originAgentCluster());
+  }
+
+  const xPermittedCrossDomainPoliciesArgs = getArgs(
+    options.permittedCrossDomainPolicies
+  );
+  if (xPermittedCrossDomainPoliciesArgs) {
+    result.push(
+      xPermittedCrossDomainPolicies(...xPermittedCrossDomainPoliciesArgs)
+    );
+  }
+
+  const referrerPolicyArgs = getArgs(options.referrerPolicy);
+  if (referrerPolicyArgs) {
+    result.push(referrerPolicy(...referrerPolicyArgs));
+  }
+
+  const xXssProtectionArgs = getArgs(options.xssFilter, {
+    name: "xssFilter",
+    takesOptions: false,
+  });
+  if (xXssProtectionArgs) {
+    result.push(xXssProtection());
+  }
+
+  return result;
 }
 
 const helmet: Helmet = Object.assign(
@@ -82,81 +224,7 @@ const helmet: Helmet = Object.assign(
       );
     }
 
-    if (
-      Object.entries(options).some(
-        ([key, option]) => option === true && key !== "originAgentCluster"
-      )
-    ) {
-      throw new Error(
-        "Helmet no longer supports `true` as a middleware option, except for Origin-Agent-Cluster. Remove the property from your options to fix this error."
-      );
-    }
-
-    const middlewareFunctions: MiddlewareFunction[] = [];
-
-    if (options.contentSecurityPolicy !== false) {
-      middlewareFunctions.push(
-        contentSecurityPolicy(options.contentSecurityPolicy)
-      );
-    }
-    if (options.dnsPrefetchControl !== false) {
-      middlewareFunctions.push(xDnsPrefetchControl(options.dnsPrefetchControl));
-    }
-    if (options.expectCt !== false) {
-      middlewareFunctions.push(expectCt(options.expectCt));
-    }
-    if (options.frameguard !== false) {
-      middlewareFunctions.push(xFrameOptions(options.frameguard));
-    }
-    if (options.hidePoweredBy !== false) {
-      if (options.hidePoweredBy !== undefined) {
-        console.warn(
-          "hidePoweredBy does not take options. Remove the property to silence this warning."
-        );
-      }
-      middlewareFunctions.push(xPoweredBy());
-    }
-    if (options.hsts !== false) {
-      middlewareFunctions.push(strictTransportSecurity(options.hsts));
-    }
-    if (options.ieNoOpen !== false) {
-      if (options.ieNoOpen !== undefined) {
-        console.warn(
-          "ieNoOpen does not take options. Remove the property to silence this warning."
-        );
-      }
-      middlewareFunctions.push(xDownloadOptions());
-    }
-    if (options.noSniff !== false) {
-      if (options.noSniff !== undefined) {
-        console.warn(
-          "noSniff does not take options. Remove the property to silence this warning."
-        );
-      }
-      middlewareFunctions.push(xContentTypeOptions());
-    }
-    if (
-      options.originAgentCluster !== undefined &&
-      options.originAgentCluster !== false
-    ) {
-      middlewareFunctions.push(originAgentCluster());
-    }
-    if (options.permittedCrossDomainPolicies !== false) {
-      middlewareFunctions.push(
-        xPermittedCrossDomainPolicies(options.permittedCrossDomainPolicies)
-      );
-    }
-    if (options.referrerPolicy !== false) {
-      middlewareFunctions.push(referrerPolicy(options.referrerPolicy));
-    }
-    if (options.xssFilter !== false) {
-      if (options.xssFilter !== undefined) {
-        console.warn(
-          "xssFilter does not take options. Remove the property to silence this warning."
-        );
-      }
-      middlewareFunctions.push(xXssProtection());
-    }
+    const middlewareFunctions = getMiddlewareFunctionsFromOptions(options);
 
     return function helmetMiddleware(
       req: IncomingMessage,
@@ -183,6 +251,7 @@ const helmet: Helmet = Object.assign(
   },
   {
     contentSecurityPolicy,
+    crossOriginEmbedderPolicy,
     dnsPrefetchControl: xDnsPrefetchControl,
     expectCt,
     frameguard: xFrameOptions,